refactor: remove ir components
