+++ apps/dapp/next.config.js
@@ -25,16 +25,6 @@ module.exports = {
         permanent: false,
       },
       {
-        source: '/ir',
-        destination: '/straddles',
-        permanent: false,
-      },
-      {
-        source: '/ir/:path*',
-        destination: '/straddles',
-        permanent: false,
-      },
-      {
         source: '/ssov-v3/:path*',
         destination: '/ssov/:path*',
         permanent: false,
