--- apps/dapp/src/components/ir/Deposits/styles.module.scss
@@ -1,26 +1,0 @@
-.optionsTable td {
-  border: none !important;
-}
-
-.optionsTable table {
-  border-collapse: separate !important;
-  border-spacing: 0 0.5em !important;
-}
-
-.optionsTable td {
-  border: solid 1px #000;
-  border-style: solid none;
-  padding: 10px 16px;
-}
-
-.optionsTable td:first-child {
-  border-left-style: solid;
-  border-top-left-radius: 10px;
-  border-bottom-left-radius: 10px;
-}
-
-.optionsTable td:last-child {
-  border-right-style: solid;
-  border-bottom-right-radius: 10px;
-  border-top-right-radius: 10px;
-}
