--- apps/dapp/src/components/ir/Description/index.tsx
@@ -1,38 +1,0 @@
-import Box from '@mui/material/Box';
-
-import { VAULT_MAP } from 'constants/index';
-import Typography from 'components/UI/Typography';
-
-const Description = ({ poolName }: { poolName: string }) => {
-  return (
-    <Box className={'lg:w-3/4'}>
-      <Box className={'flex'}>
-        <Box className={'rounded-full mt-auto mb-auto'}>
-          <img
-            src={VAULT_MAP[poolName]?.src}
-            className="w-[6rem]"
-            alt={poolName}
-          />
-        </Box>
-        <Typography
-          variant="h3"
-          className="ml-5 flex items-center space-x-3 lg:text-4xl"
-        >
-          {poolName} Interest Rate Vault
-        </Typography>
-        <Typography
-          variant="h3"
-          className="mt-8 ml-6 mb-6 h-10 flex text-lg items-center space-x-3 border border-primary py-1 px-2 rounded-md"
-        >
-          BETA
-        </Typography>
-      </Box>
-      <Typography variant="h4" className="text-stieglitz mt-6 mb-6">
-        This vault accepts 2CRV LP deposits and lets users write Interest Rate
-        options that allows for bet/hedge on the underlying interest rate.
-      </Typography>
-    </Box>
-  );
-};
-
-export default Description;
