--- apps/dapp/src/components/ir/PurchaseCard/styles.module.scss
@@ -1,6 +1,0 @@
-.cardWidth {
-  width: 100%;
-  @media (min-width: 768px) {
-    width: 23rem;
-  }
-}
