--- apps/dapp/src/components/ir/SelectStrikeWidget/styles.module.scss
@@ -1,6 +1,0 @@
-.cardWidth {
-  width: 100%;
-  @media (min-width: 768px) {
-    width: 21rem;
-  }
-}
